# Conflicts:
#	tp_integrador/tp_integrador/bin/tp_integrador.dll
#	tp_integrador/tp_integrador/bin/tp_integrador.pdb
#	tp_integrador/tp_integrador/obj/Debug/tp_integrador.csproj.CoreCompileInputs.cache
#	tp_integrador/tp_integrador/obj/Debug/tp_integrador.csproj.FileListAbsolute.txt
#	tp_integrador/tp_integrador/obj/Debug/tp_integrador.dll
#	tp_integrador/tp_integrador/obj/Debug/tp_integrador.pdb

+++ tp_integrador/tp_integrador/Models/Cliente.cs
@@ -73,7 +73,7 @@ namespace tp_integrador.Models
             aparato.ModoAhorro();
         }
 
-        public int GetEstimado(Estandar aparato)
+        public double GetEstimado(Estandar aparato)
         {
             return aparato.ConsumoEstimado();
         }
